+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/observables/ProfileObservable.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-package com.algoritmico.passepartout.ui
+package com.algoritmico.passepartout.observables
 
 import com.algoritmico.passepartout.Globals
 import com.algoritmico.passepartout.abi.AppABIProfileProtocol
