+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/observables/UserPreferencesObservable.kt
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: GPL-3.0
 
-package com.algoritmico.passepartout.ui
+package com.algoritmico.passepartout.observables
 
 import android.content.Context
 import android.util.Log
