+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/AppCoordinator.kt
@@ -36,6 +36,9 @@ import androidx.compose.ui.graphics.vector.path
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.algoritmico.passepartout.Globals
+import com.algoritmico.passepartout.observables.ProfileObservable
+import com.algoritmico.passepartout.observables.TunnelObservable
+import com.algoritmico.passepartout.observables.UserPreferencesObservable
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
