+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/PassepartoutApp.kt
@@ -12,6 +12,9 @@ import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import com.algoritmico.passepartout.observables.ProfileObservable
+import com.algoritmico.passepartout.observables.TunnelObservable
+import com.algoritmico.passepartout.observables.UserPreferencesObservable
 
 @Composable
 fun PassepartoutApp(
