+++ app-android/app/src/main/kotlin/com/algoritmico/passepartout/ui/SettingsCoordinator.kt
@@ -38,6 +38,7 @@ import androidx.compose.ui.graphics.vector.path
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
+import com.algoritmico.passepartout.observables.UserPreferencesObservable
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
